# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopWave E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: brand
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: minRating
          in: query
          schema: { type: number }
        - name: inStock
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, newest, bestseller, rating]
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (seller/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (seller/admin)
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (seller/admin)
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── CATEGORIES ────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Post a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── ADDRESSES ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ProductVariant:
      type: object
      properties:
        name: { type: string }
        value: { type: string }
        priceModifier: { type: number }
        stock: { type: integer }
      required: [name, value, stock]
    Product:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        categoryName: { type: string }
        brand: { type: string }
        price: { type: number }
        discountPercent: { type: number }
        finalPrice: { type: number }
        stock: { type: integer }
        sellerId: { type: integer }
        sellerName: { type: string }
        avgRating: { type: number }
        reviewCount: { type: integer }
        variants: { type: array, items: { $ref: "#/components/schemas/ProductVariant" } }
        isFeatured: { type: boolean }
        isFlashSale: { type: boolean }
        flashSaleEnd: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, title, slug, price, stock, images, categoryId, brand, avgRating, reviewCount, finalPrice, discountPercent, sellerId, sellerName, isFeatured, isFlashSale, createdAt]
    ProductInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        brand: { type: string }
        price: { type: number }
        discountPercent: { type: number }
        stock: { type: integer }
        variants: { type: array, items: { $ref: "#/components/schemas/ProductVariant" } }
        isFeatured: { type: boolean }
        isFlashSale: { type: boolean }
        flashSaleEnd: { type: ["string", "null"] }
      required: [title, description, images, categoryId, brand, price, stock]
    ProductUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        brand: { type: string }
        price: { type: number }
        discountPercent: { type: number }
        stock: { type: integer }
        isFeatured: { type: boolean }
        isFlashSale: { type: boolean }
        flashSaleEnd: { type: ["string", "null"] }
    ProductList:
      type: object
      properties:
        products: { type: array, items: { $ref: "#/components/schemas/Product" } }
        total: { type: integer }
        page: { type: integer }
        totalPages: { type: integer }
      required: [products, total, page, totalPages]
...
    Review:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        productId: { type: integer }
        rating: { type: integer }
        title: { type: string }
        body: { type: string }
        helpfulVotes: { type: integer }
        createdAt: { type: string }
      required: [id, userId, userName, productId, rating, title, body, helpfulVotes, createdAt]
    ReviewInput:
      type: object
      properties:
        rating: { type: integer }
        title: { type: string }
        body: { type: string }
      required: [rating, title, body]

    # ─── ADDRESS ──────────────────────────────────────────────────────────────
...