# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariant } from './productVariant';

export interface ProductInput {
  title: string;
  description: string;
  images: string[];
  categoryId: number;
  brand: string;
  price: number;
  discountPercent?: number;
  stock: number;
  variants?: ProductVariant[];
  isFeatured?: boolean;
  isFlashSale?: boolean;
  /** @nullable */
  flashSaleEnd?: string | null;
}